added note about creating openrc

+++ install.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [[ ! -e ./openrc.sh ]]; then
+  echo "NO OPENRC FOUND! CREATE ONE, AND TRY AGAIN!"
+  exit
+fi
+
 yum -y install https://github.com/openhpc/ohpc/releases/download/v1.3.GA/ohpc-release-1.3-1.el7.x86_64.rpm
 
 yum -y install ohpc-slurm-server ansible
